Move IngressRoute generated client to apis package
