+++ hack/update-generated-crd-code.sh
@@ -20,7 +20,7 @@ REPO_ROOT=${HACK_DIR}/..
 
 ${REPO_ROOT}/vendor/k8s.io/code-generator/generate-groups.sh \
   all \
-  github.com/heptio/contour/internal/generated \
+  github.com/heptio/contour/apis/generated \
   github.com/heptio/contour/apis \
   contour:v1beta1 \
   --go-header-file hack/boilerplate.go.tmpl \
