+++ internal/k8s/status_test.go
@@ -18,7 +18,7 @@ import (
 	"testing"
 
 	ingressroutev1beta1 "github.com/heptio/contour/apis/contour/v1beta1"
-	"github.com/heptio/contour/internal/generated/clientset/versioned/fake"
+	"github.com/heptio/contour/apis/generated/clientset/versioned/fake"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	k8stesting "k8s.io/client-go/testing"
