+++ internal/k8s/watcher.go
@@ -18,7 +18,7 @@ import (
 	"time"
 
 	ingressroutev1 "github.com/heptio/contour/apis/contour/v1beta1"
-	clientset "github.com/heptio/contour/internal/generated/clientset/versioned"
+	clientset "github.com/heptio/contour/apis/generated/clientset/versioned"
 	"github.com/heptio/workgroup"
 	"github.com/sirupsen/logrus"
